inform-retagify: match objects declared with class name
Replaces the regex used to call the tags program with a more general one
that does not rely on Object/Nearby/Class/[ at the start of a
definition. This allows objects declared with the syntax 'classname
objectname' to be matched. It will also match other directives eg Array.

Matching other directives may cause problems, however: for this example

Replace XSub;
...
[ XSub;

the first instance of XSub will get the tag when it should be the second.

Change based on a patch by Marshall T. Vandegrift.